upload Dalibor's PubMed abstract miner
This script was written for a specific author: Pirnik. It is based on an abstract miner which I wrote at the beginning of my PhD in the UMCG to get a quick overview of publications of fellow colleagues.

It uses specific packages to query Medline for the given author, select only specific components of a publication and boil the text down into an .Rmd file. This file is finally rendered into an html file for neat viewing. The script also creates a wordcloud with the most frequent words from the authors abstracts.