Create new downloadscript
We could iteratively load the results so that the first  result pops up immediately and then a second later the second one comes? I'm having some difficulties making it faster... It seems  This solves the bug also where RSImed somehow does not have results even though there is data on the pubmed site.